requirements.txt
playwright==1.40.0
anthropic==0.25.0
python-dotenv==1.0.0
pydantic==2.5.0
